feat: create Home page
