--- src/components/Main/Main.tsx
@@ -1,89 +1,0 @@
-import { ReactElement, useEffect, useState } from "react";
-import { useSearchParams } from "react-router-dom";
-import { SearchInput } from "../SearchInput";
-import { Button } from "../Button";
-import { Message } from "../Message";
-import { SearchResults } from "../SearchResults";
-import { Pagination } from "../Pagination";
-import { useFetching } from "../../hooks/useFetching";
-import { LIMIT } from "../../constants";
-
-import classnames from "classnames";
-import styles from "./main.module.scss";
-import { Details } from "../Details";
-
-// Limit page and limit
-// Change strings to numbers
-// Remake pagination w next, prev, first, last
-// Main to Home
-// Split into 2 components: Results and Details
-
-function Main(): ReactElement {
-  const [fetchData, searchResults, isLoading, isError, setIsError] = useFetching();
-
-  const [searchParams, setSearchParams] = useSearchParams("");
-
-  const [page, setPage] = useState(searchParams.get("offset") || "1");
-  const [limit, setLimit] = useState(searchParams.get("limit") || LIMIT);
-
-  const [selectedItem, setSelectedItem] = useState<null | string>(null);
-
-  useEffect(() => {
-    setPage("1");
-    setSearchParams({ limit: limit, offset: page });
-  }, [limit]);
-
-  useEffect(() => {
-    setSearchParams({ limit: limit, offset: page });
-  }, [page]);
-
-  useEffect(() => {
-    const limitURL = searchParams.get("limit") || LIMIT;
-
-    setSearchParams({ limit: limitURL, offset: page });
-    fetchData(searchParams.toString());
-  }, [searchParams]);
-
-  useEffect(() => {
-    if (isError) {
-      throw new Error("Test error is thrown!");
-    }
-  }, [isError]);
-
-  function handleSearch(searchString: string): void {
-    fetchData(searchString);
-  }
-
-  function handleThrowError(): void {
-    setIsError(true);
-  }
-
-  return (
-    <div className={classnames(styles.page, selectedItem ? styles.page__split : "")}>
-      <div className={classnames(styles.page__column, styles.page__results)}>
-        <header className="header">
-          <span className="note">The total number of Pokémon is limited to 300 due to practical considerations.</span>
-
-          <SearchInput onSearch={handleSearch} isLoading={isLoading} />
-          <Button title="Throw Error ⚡️" onClick={handleThrowError} disabled={isLoading} />
-        </header>
-
-        {isError && <Message message="Oops!.. Something wrong. Try again!" />}
-        {isLoading && <Message message="Loading..." />}
-
-        {!isError && !isLoading && (
-          <SearchResults searchResults={searchResults} page={page} limit={limit} setSelectedItem={setSelectedItem} />
-        )}
-
-        <Pagination page={page} setPage={setPage} limit={limit} setLimit={setLimit} isLoading={isLoading} />
-      </div>
-      {selectedItem && (
-        <div className={classnames(styles.page__column, styles.page__details)}>
-          <Details selectedItem={selectedItem} setSelectedItem={setSelectedItem} />
-        </div>
-      )}
-    </div>
-  );
-}
-
-export default Main;
