--- src/components/Main/index.ts
@@ -1,2 +1,0 @@
-import Main from "./Main";
-export { Main };
