+++ src/pages/Home/index.ts
@@ -1,0 +1,2 @@
+import Home from "./Home";
+export { Home };
